remove ZipEntryOutputStream
