+++ lib/src/main/java/io/reist/sklad/ZipStorage.java
@@ -3,11 +3,16 @@ package io.reist.sklad;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
+import java.io.BufferedOutputStream;
 import java.io.File;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.util.HashSet;
+import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
+import java.util.zip.ZipOutputStream;
 
 import io.reist.sklad.utils.FileUtils;
 import io.reist.sklad.utils.IOUtils;
@@ -50,7 +55,47 @@ public class ZipStorage implements Storage {
     @NonNull
     @Override
     public OutputStream openOutputStream(@NonNull String id) throws IOException {
-        return new ZipEntryOutputStream(file, id);
+
+        final File tmpFile = FileUtils.tempFile(new File(file.getParent()));
+        final ZipOutputStream out = new ZipOutputStream(new BufferedOutputStream(new FileOutputStream(tmpFile)));
+
+        HashSet<String> skipSet = new HashSet<>();
+        skipSet.add(id);
+        ZipUtils.copyEntries(file, out, skipSet);
+
+        out.putNextEntry(new ZipEntry(id));
+
+        return new OutputStream() {
+            @Override
+            public void write(int b) throws IOException {
+                out.write(b);
+            }
+
+            @Override
+            public void close() throws IOException {
+                out.closeEntry();
+                out.close();
+
+                FileUtils.deleteFile(file);
+                FileUtils.moveFile(tmpFile, file);
+                FileUtils.deleteFile(tmpFile);
+            }
+
+            @Override
+            public void write(byte[] b) throws IOException {
+                out.write(b);
+            }
+
+            @Override
+            public void flush() throws IOException {
+                out.flush();
+            }
+
+            @Override
+            public void write(@NonNull byte[] b, int off, int len) throws IOException {
+                out.write(b, off, len);
+            }
+        };
     }
 
     @Override
